Added a fix for 64 bits windows. The type of the SecHandle structure was hardcoded to int 32 bit int values which was giving problems on 64 bits windows. Now uses an architecture independente pointer type. Thanks Brar Piening for patch and heads up.

+++ src/Npgsql/SSPIHandler.cs
@@ -25,8 +25,8 @@ namespace Npgsql
         [StructLayout(LayoutKind.Sequential)]
         private struct SecHandle
         {
-            public int dwLower;
-            public int dwUpper;
+            public IntPtr dwLower;
+            public IntPtr dwUpper;
         }
 
         [StructLayout(LayoutKind.Sequential)]
